Move css files to separate directory.
