+++ client/src/css/VideoPlayerLarge.css
@@ -1,3 +1,7 @@
+html {
+  scroll-behavior: smooth;
+}
+
 html, body {
   height: 100%;
   width: 100%;
